modularise steps and make instructor register importable component
